backend/requirements.txt
# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django-environ>=0.11,<1.0
django-extensions>=3.2,<4.0
django-redis>=5.4,<6.0
django-storages>=1.14,<2.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Task Queue
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# Security
argon2-cffi>=23.1,<24.0
python-decouple>=3.8,<4.0

# Image Processing
Pillow>=10.0,<11.0

# API Documentation
drf-spectacular>=0.27,<1.0

# AWS S3
boto3>=1.34,<2.0

# Sentry (Error Tracking)
sentry-sdk>=1.39,<2.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<23.0

# Development
black>=24.0,<25.0
flake8>=7.0,<8.0
isort>=5.13,<6.0
django-debug-toolbar>=4.2,<5.0

# Production
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0

# Utilities
python-slugify>=8.0,<9.0
